requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
docutils==0.21.2
filetype==1.2.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_matplotlib_widget==0.13.1
kivymd==1.2.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
psutil==7.0.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
scipy==1.15.2
six==1.17.0
smmap==5.0.2
urllib3==2.3.0
